hello

+++ src/main/java/com/ra/controller/admin/CategoryController.java
@@ -1,6 +1,6 @@
 package com.ra.controller.admin;
 
-import com.ra.exception.CustomException;
+import com.ra.exception.CategoryException;
 import com.ra.model.dto.request.CategoryRequestDTO;
 import com.ra.model.dto.response.CategoryResponseDTO;
 import com.ra.model.entity.Category;
@@ -15,6 +15,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 @RestController
+@CrossOrigin("*")
 @RequestMapping("/admin")
 public class CategoryController {
     @Autowired
@@ -23,7 +24,7 @@ public class CategoryController {
 
     // sort-pagination
     @GetMapping("/category/sort-pagination")
-    public ResponseEntity<Page<Category>> categoryIndex(
+    public ResponseEntity<?> categoryIndex(
             @RequestParam(name = "page", defaultValue = "0") int page,
             @RequestParam(name = "size", defaultValue = "3") int size,
             @RequestParam(name = "sort", defaultValue = "id") String sort,
@@ -40,12 +41,12 @@ public class CategoryController {
 
     // search
     @GetMapping("/category/search")
-    public ResponseEntity<Page<CategoryResponseDTO>> categorySearch(
+    public ResponseEntity<?> categorySearch(
             @RequestParam(name = "page", defaultValue = "0") int page,
             @RequestParam(name = "size", defaultValue = "3") int size,
             @RequestParam(name = "sort", defaultValue = "id") String sort,
             @RequestParam(name = "order", defaultValue = "asc") String order,
-            @RequestParam(name = "search") String search) {
+            @RequestParam(name = "search") String search) throws CategoryException {
         Pageable pageable;
         if (order.equals("asc")) {
             pageable = PageRequest.of(page, size, Sort.by(sort).ascending());
@@ -58,7 +59,7 @@ public class CategoryController {
 
     // add
     @PostMapping("/category")
-    public ResponseEntity<CategoryResponseDTO> addCategory(@RequestBody CategoryRequestDTO categoryDTO, Long id) throws CustomException {
+    public ResponseEntity<CategoryResponseDTO> addCategory(@RequestBody CategoryRequestDTO categoryDTO, Long id) throws CategoryException {
         CategoryResponseDTO newCategory = categoryService.saveOrUpdate(categoryDTO);
         return new ResponseEntity<>(newCategory, HttpStatus.CREATED);
     }
@@ -75,8 +76,11 @@ public class CategoryController {
 
     // update
     @PutMapping("/category/{id}")
-    public ResponseEntity<CategoryResponseDTO> updateCategory(@PathVariable("id") Long id, @RequestBody CategoryRequestDTO categoryDTO) throws CustomException {
+    public ResponseEntity<?> updateCategory(@PathVariable("id") Long id, @ModelAttribute CategoryRequestDTO categoryDTO) throws CategoryException {
         CategoryResponseDTO category = categoryService.findById(id);
+        if (category == null) {
+            return new ResponseEntity<>("Could not find the id of the category that needs repair", HttpStatus.BAD_REQUEST);
+        }
         categoryDTO.setId(category.getId());
         CategoryResponseDTO updatedCategory = categoryService.saveOrUpdate(categoryDTO);
         return new ResponseEntity<>(updatedCategory, HttpStatus.OK);
@@ -85,18 +89,26 @@ public class CategoryController {
     // change status
     @PatchMapping("/category/{id}")
     public ResponseEntity<?> changeStatus(@PathVariable Long id) {
-        categoryService.changeStatus(id);
         CategoryResponseDTO categoryResponseDTO = categoryService.findById(id);
-        return new ResponseEntity<>(categoryResponseDTO, HttpStatus.OK);
+        if (categoryResponseDTO == null) {
+            return new ResponseEntity<>("Could not find the id of the category that needs repair", HttpStatus.BAD_REQUEST);
+        }
+        categoryService.changeStatus(id);
+        return new ResponseEntity<>("Status change successful", HttpStatus.OK);
     }
 
     // delete
     @DeleteMapping("/category/{id}")
-    public ResponseEntity<Void> deleteCategory(@PathVariable("id") Long id) {
+    public ResponseEntity<?> deleteCategory(@PathVariable("id") Long id) {
         CategoryResponseDTO category = categoryService.findById(id);
         if (category != null) {
+            // Kiểm tra xem danh mục có sản phẩm hay không
+            boolean checkProductsInCategory = categoryService.checkProductsInCategory(id);
+            if (checkProductsInCategory) {
+                return new ResponseEntity<>("This category contains products that cannot be deleted", HttpStatus.BAD_REQUEST);
+            }
             categoryService.delete(id);
-            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+            return new ResponseEntity<>("deleted successfully", HttpStatus.OK);
         }
         return new ResponseEntity<>(HttpStatus.NOT_FOUND);
     }
